+++ libs/eco-store/cart/feature/src/lib/eco-store-cart-steps/shipping/form/cart-shipping-form.config.ts
@@ -64,7 +64,7 @@ export function getCartShippingFormConfig(): FormConfig<EcoStoreCartState> {
     tenantStore,
     cartStore,
     translateService,
-    tenantAddresses: tenantStore.getTenantAddressesContacts(),
+    tenantAddresses: tenantStore.tenantAddressesContacts(),
     userAddresses: userProfileStore.getUserContacts(),
     availableMethodTypes,
   };