+++ libs/eco-store/core/entities/src/tenant.ts
@@ -9,6 +9,16 @@ export type SlotDays =
   | 'saturday'
   | 'sunday';
 
+export const DAYS_MAP: Record<SlotDays, number> = {
+  monday: 1,
+  tuesday: 2,
+  wednesday: 3,
+  thursday: 4,
+  friday: 5,
+  saturday: 6,
+  sunday: 0,
+} as const;
+
 export type TimePoint = string;
 export type TimeRange = string;
 
@@ -20,7 +30,7 @@ export interface EcoStoreTenantLogisticsDeliveryTier {
 
 export interface EcoStoreTenantLogisticsDeliveryOption {
   type: EcoStoreTenantLogisticsDeliveryType;
-  enabled: boolean;
+  enabled: boolean; // enable or disable any access to the tenant store.
   cost?: number; // you can set a fixed cost or a percentage of the order total. For more complex pricing rules you can use the slot price rules.
   addressOverride?: string; // if set, this address will be used instead of the tenant address
   instructions?: string | LocalizedFields<string> | null; // instructions for the delivery/pickup
@@ -45,7 +55,9 @@ export type EcoStoreTenant = BasePocketBaseEntity &
     languages: string[];
     themeConfig?: unknown;
     logisticsConfig: EcoStoreTenantLogistics;
-    active?: boolean;
+    active?: boolean; // closing at superuser level.
+    closed?: boolean; // closing at tenant level.
+    closedReason?: LocalizedFields<string> | null; // closed by tenant reason.
   };
 
 export type EcoStoreTenantAddress = Pick<
@@ -56,3 +68,17 @@ export type EcoStoreTenantAddress = Pick<
     tenant: EcoStoreTenant['id'];
     slots: Record<SlotDays, TimeRange[]>;
   };
+
+/*
+- OPEN - currently open.
+- CLOSED - currently closed. Means it is closed because the tenant window is just open some days in a week or some hours in a day but open in the future.
+- OPENING_SOON - currently closed, but it will open soon. Just an intermediate state between CLOSED and OPEN to allow to make UI styling and message different.
+- CLOSED_MANUALLY - currently closed by the tenant, and it can have a message for the reasons.
+- CANCELLED - currently closed by the superuser.
+*/
+export type EcoStoreTenantWindowStatus =
+  | 'OPEN'
+  | 'CLOSED'
+  | 'OPENING_SOON'
+  | 'CLOSED_MANUALLY'
+  | 'CANCELLED';