+++ libs/eco-store/core/layout/src/header/header.component.ts
@@ -5,8 +5,9 @@ import { MatToolbar } from '@angular/material/toolbar';
 import { RouterLink } from '@angular/router';
 import { TranslateModule } from '@ngx-translate/core';
 import { FormConfig } from '@plastik/core/entities';
-import { EcoStoreTenant } from '@plastik/eco-store/entities';
+import { EcoStoreTenant, EcoStoreTenantWindowStatus } from '@plastik/eco-store/entities';
 import { PocketBaseImageUrlPipe } from '@plastik/eco-store/shared/utils';
+import { StoreWindowComponent } from '@plastik/eco-store/store-window';
 import { SharedFormFeatureComponent } from '@plastik/shared/form';
 import { SharedImgContainerComponent } from '@plastik/shared/img-container';
 
@@ -20,6 +21,7 @@ import { SharedImgContainerComponent } from '@plastik/shared/img-container';
     TranslateModule,
     SharedFormFeatureComponent,
     SharedImgContainerComponent,
+    StoreWindowComponent,
     PocketBaseImageUrlPipe,
   ],
   templateUrl: './header.component.html',
@@ -29,6 +31,10 @@ import { SharedImgContainerComponent } from '@plastik/shared/img-container';
 export class HeaderComponent {
   readonly formConfig = input<FormConfig<{ query: string }>>();
   readonly tenant = input<EcoStoreTenant | null>();
+  readonly storeStatus = input<EcoStoreTenantWindowStatus | undefined>();
+  readonly nextOpenDate = input<Date | null | undefined>();
+  readonly is24h = input<boolean | undefined>();
+  readonly closedReason = input<string | null | undefined>();
   readonly submitEvent = output<{ query: string }>();
 
   protected onSubmit(event: { query: string }): void {