+++ libs/eco-store/store-window/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/store-window/store-window.component';